move SQLErrorCode from sharding-proxy-common to sharding-proxy-transport-core module
