+++ sharding-proxy/sharding-proxy-backend/pom.xml
@@ -29,6 +29,11 @@
     <dependencies>
         <dependency>
             <groupId>org.apache.shardingsphere</groupId>
+            <artifactId>sharding-proxy-transport-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.shardingsphere</groupId>
             <artifactId>sharding-proxy-common</artifactId>
             <version>${project.version}</version>
         </dependency>
