+++ sharding-proxy/sharding-proxy-transport/sharding-proxy-transport-core/src/main/java/org/apache/shardingsphere/shardingproxy/transport/error/SQLErrorCode.java
@@ -15,27 +15,31 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.shardingproxy.error;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
+package org.apache.shardingsphere.shardingproxy.transport.error;
 
 /**
- * Common error code.
+ * SQL error code.
  */
-@RequiredArgsConstructor
-@Getter
-public enum CommonErrorCode implements SQLErrorCode {
+public interface SQLErrorCode {
     
-    CIRCUIT_BREAK_MODE(10000, "C10000", "Circuit break mode is ON."),
+    /**
+     * Get error code.
+     * 
+     * @return error code
+     */
+    int getErrorCode();
     
-    UNSUPPORTED_COMMAND(10001, "C10001", "Unsupported command: [%s]"),
+    /**
+     * Get SQL state.
+     * 
+     * @return SQL state
+     */
+    String getSqlState();
     
-    UNKNOWN_EXCEPTION(10002, "C10002", "Unknown exception: [%s]");
-    
-    private final int errorCode;
-    
-    private final String sqlState;
-    
-    private final String errorMessage;
+    /**
+     * Get error message.
+     * 
+     * @return error message
+     */
+    String getErrorMessage();
 }
