+++ sharding-proxy/sharding-proxy-transport/sharding-proxy-transport-core/src/test/java/org/apache/shardingsphere/shardingproxy/transport/error/CommonErrorCodeTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.shardingproxy.error;
+package org.apache.shardingsphere.shardingproxy.transport.error;
 
 import org.junit.Test;
