+++ sharding-proxy/sharding-proxy-transport/sharding-proxy-transport-mysql/src/main/java/org/apache/shardingsphere/shardingproxy/transport/mysql/constant/MySQLServerErrorCode.java
@@ -19,7 +19,7 @@ package org.apache.shardingsphere.shardingproxy.transport.mysql.constant;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.shardingproxy.error.SQLErrorCode;
+import org.apache.shardingsphere.shardingproxy.transport.error.SQLErrorCode;
 
 /**
  * Server error code for MySQL.
